Move data model files to seperate folder.
* Combined all related user code (db model, manager, and exceptions) into user.py
* Moved community resource code into community_resource.py
* Renamed files, removed _models suffix to make it less verbose.
* Fixed imports accordingly.
