+++ app/models/user.py
@@ -2,8 +2,10 @@ from flask import current_app
 from itsdangerous import (TimedJSONWebSignatureSerializer
                           as Serializer, BadSignature, SignatureExpired)
 from werkzeug.security import check_password_hash, generate_password_hash
+from ..validators import is_valid_username
 
-from . import db
+from .. import db
+
 
 class User(db.Model):
     __tablename__ = "users"
@@ -12,27 +14,23 @@ class User(db.Model):
     username = db.Column(db.String(64), unique=True,
                          nullable=False, index=True)
     password_hash = db.Column(db.String(256), nullable=False)
-    #email = db.Column(db.String(64), unique=True,
+    # email = db.Column(db.String(64), unique=True,
     #                    nullable=True, index=True)
 
     @property
     def password(self):
         raise AttributeError('password is not a readable attribute')
 
-
     @password.setter
     def password(self, password):
         self.password_hash = generate_password_hash(password)
 
-
     def verify_password(self, password):
         return check_password_hash(self.password_hash, password)
-
 
     def generate_auth_token(self, expiration=600):
         s = Serializer(current_app.config['SECRET_KEY'], expires_in=expiration)
         return s.dumps({'id': self.id})
-
 
     @staticmethod
     def verify_auth_token(token):
@@ -46,18 +44,15 @@ class User(db.Model):
         user = User.query.get(data['id'])
         return user
 
-
     def to_dict(self):
         return {
             "id": self.id,
             "username": self.username
         }
 
-
     @staticmethod
     def get_user_by_username(username):
         return User.query.filter_by(username=username).first()
-
 
     @staticmethod
     def add_user(user):
@@ -67,3 +62,30 @@ class User(db.Model):
         db.session.add(user)
         db.session.commit()
         return user
+
+
+class UserManager():
+
+    @staticmethod
+    def edit_user(user_id, new_username):
+        user = User.query.get(user_id)
+        try:
+            if user is None:
+                raise NoExistingUser("Something went wrong!")
+            if not is_valid_username(new_username):
+                raise InvalidUserInfo("User information is invalid.")
+            user.username = new_username
+        except NoExistingUser:
+            raise
+        except InvalidUserInfo:
+            raise
+
+
+class NoExistingUser(Exception):
+    def __init__(self, message):
+        Exception.__init__(self, message)
+
+
+class InvalidUserInfo(Exception):
+    def __init__(self, message):
+        Exception.__init__(self, message)
